requirements.txt
django
django-paintstore
